requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
langextract[openai]>=0.1.0
python-dotenv>=1.0
google-genai>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9
dj-database-url>=2.1
